// frontend/src/lib/api/spec.ts
export interface components {
    schemas: {
        /** @description This is a BackupJobState union type. It is used to represent the state of a backup job. */
        BackupJobState: components["schemas"]["BackupJobState_IncrementalBackupState"];
        /** @description This is a BackupJobState union type. It is used to represent the state of a backup job. */
        BackupJobState_IncrementalBackupState: {
            /**
             * @example Incremental
             * @enum {string}
             */
            type: "Incremental";
        } & components["schemas"]["IncrementalBackupState"];
        /** @enum {string} */
        CompressionLevel: "None" | "Fast" | "Balanced" | "Best";
        /** FetchingMetadataState */
        FetchingMetadataState: Record<string, never>;
        IncrementalBackupStage: components["schemas"]["IncrementalBackupStage_FetchingMetadataState"] | components["schemas"]["IncrementalBackupStage_IncrementalBackupUploadState"];
        IncrementalBackupStage_FetchingMetadataState: {
            /**
             * @example FetchingMetadata
             * @enum {string}
             */
            stage: "FetchingMetadata";
        } & components["schemas"]["FetchingMetadataState"];
        IncrementalBackupStage_IncrementalBackupUploadState: {
            /**
             * @example Uploading
             * @enum {string}
             */
            stage: "Uploading";
        } & components["schemas"]["IncrementalBackupUploadState"];
        /** IncrementalBackupState */
        IncrementalBackupState: {
            /** Format: date-time */
            started_at: string;
            stage: components["schemas"]["IncrementalBackupStage"];
        };
        /** IncrementalBackupUploadState */
        IncrementalBackupUploadState: {
            /** Format: date-time */
            timestamp: string;
            /** Format: uint32 */
            parent?: number;
            remote_filename: string;
            local_snapshot: string;
            /** Format: uint64 */
            bytes_read: number;
            /** Format: uint64 */
            bytes_written: number;
            compression_level: components["schemas"]["CompressionLevel"];
            encrypted: boolean;
            finishing: boolean;
        };
        /** JobStates */
        JobStates: {
            /** @description Contains the state of the restore job if it is running. */
            restore?: components["schemas"]["RestoreJobState"] & unknown;
            /** @description Contains the state of the backup job if it is running. */
            backup?: Omit<components["schemas"]["BackupJobState"], "type"> & unknown;
        };
        /** RestoreJobState */
        RestoreJobState: Record<string, never>;
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}